requirements.txt
# FastAPI & Web Framework
fastapi==0.121.0
uvicorn==0.38.0
starlette==0.49.3
pydantic==2.12.3
python-multipart==0.0.20

# MQTT
paho-mqtt==2.1.0

# InfluxDB
influxdb-client==1.49.0

# Environment & Utilities
python-dotenv==1.2.1
python-dateutil==2.9.0.post0

# HTTP Client
aiohttp==3.13.2
httpx==0.28.1
requests==2.32.5

# LLM & AI (Optional - for LLM features)
openai==2.6.1

# Common Dependencies
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
urllib3==2.5.0
anyio==4.11.0
h11==0.16.0
httpcore==1.0.9
sniffio==1.3.1
typing_extensions==4.15.0

# PostgreSQL async/sync drivers and timezone data
asyncpg==0.27.0
psycopg2-binary==2.9.6
tzdata
